# backend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "< furniro - Shop APIS ❤️💻 >",
    "description": "",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "User token for authentication: **Bearer [user_token]**"
      }
    }
  },
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "summary": "Get New Access Token",
        "description": "Retrieves a new access token using the refresh token stored in cookies.",
        "operationId": "getNewAccessToken",
        "tags": ["Auth 🔐"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "example": "djhdi-dehdhi-ajeucm-alcsaxclamxj"
                  }
                },
                "required": ["refreshToken"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New access and Refreshtoken generated successfully.",
            "content": {
              "application/json": {
                "example": {
                  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "AccessToken-expiresIn": "14 minute",
                  "refreshToken": "dcbdc-kjcjdsc-cjdc-sjjdsc",
                  "RefreshToken-expiresIn": "7 day"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token not provided or invalid."
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
  }
}